Remove emote_utils/ directory
- This directory was only used to store exception and a single class
  that is unacceptable since this class wasn't even referenced anywhere
  outside of Emotes cog, so there is no point in having a separate file
  for it.
