--- bot/cogs/emote_utils/exceptions.py
@@ -1,6 +1,0 @@
-class InvalidCommandException(Exception):
-    pass
-
-
-class EmoteNotFoundException(Exception):
-    pass
